@@ -2,21 +2,19 @@
   "private": true,
   "name": "nr1-nerdpack-innovation-cicd",
   "description": "Nerdpack innovation-cicd",
-  "version": "1.4.7",
+  "version": "1.5.0",
   "scripts": {
     "start": "nr1 nerdpack:serve",
     "test": "exit 0",
     "eslint-check": "eslint nerdlets/",
     "eslint-fix": "eslint nerdlets/ --fix"
   },
-  "bugs": {
-    "email": "braj@newrelic.com"
-  },
   "dependencies": {
+    "dayjs": "^1.10.4",
     "prop-types": "^15.6.2",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
-    "semantic-ui-react": "^0.88.2"
+    "react-modal": "^3.13.1"
   },
   "browserslist": [
     "last 2 versions",
@@ -24,10 +22,14 @@
     "not dead"
   ],
   "devDependencies": {
-    "@newrelic/eslint-plugin-newrelic": "^0.3.0",
-    "@semantic-release/changelog": "^3.0.5",
-    "@semantic-release/git": "^7.0.18",
-    "eslint": "^6.6.0",
-    "prettier": "^1.19.1"
+    "@newrelic/eslint-plugin-newrelic": "^0.3.1",
+    "@semantic-release/changelog": "^5.0.1",
+    "@semantic-release/git": "^9.0.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.15.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-react": "^7.23.2",
+    "prettier": "1.19.1"
   }
 }
