@@ -1,31 +1,30 @@
-import React from 'react';
+import React from 'react'
 import {
-  Grid,
-  GridItem,
-  Stack,
-  StackItem,
   NrqlQuery,
   NerdGraphQuery,
   Spinner,
-  HeadingText,
   PlatformStateContext,
   nerdlet,
-  UserStorageQuery,
-  Radio,
-} from 'nr1';
-import ProjectDropdown from './project-dropdown';
-import SprintList from './sprint-list';
-import Dashboard from './dashboard';
-import Mapping from './mapping';
-const gqlQuery = require('./query');
+  Layout,
+  LayoutItem,
+  SegmentedControl,
+  SegmentedControlItem,
+  navigation,
+  Button,
+} from 'nr1'
+import ProjectDropdown from './toolbar/project-dropdown'
+import Navigation from './nav/navigation'
+import Dashboard from './dashboards/dashboard-container'
+import ConfigModal from './config-modal'
+import { NerdstoreDefaults, readUserCollection } from './data/nerdstore'
+const gqlQuery = require('./query')
 
 export default class InnovationCICD extends React.Component {
-  static contextType = PlatformStateContext;
+  static contextType = PlatformStateContext
 
   constructor(props) {
-    super(props);
+    super(props)
     this.state = {
-      accountId: 1382490,
       globalTime: '90 days',
       loading: true,
       eventMissing: false,
@@ -43,11 +42,11 @@ export default class InnovationCICD extends React.Component {
       issueTypeSelected: [],
       codeRepoSelected: [],
       kanbanOnly: false,
-    };
-    this.handleProjectChange = this.handleProjectChange.bind(this);
-    this.handleSprintChange = this.handleSprintChange.bind(this);
-    this.handleSettingsChange = this.handleSettingsChange.bind(this);
-    this.handleKanbanChange = this.handleKanbanChange.bind(this);
+    }
+    this.handleProjectChange = this.handleProjectChange.bind(this)
+    this.handleSprintChange = this.handleSprintChange.bind(this)
+    this.handleSettingsChange = this.handleSettingsChange.bind(this)
+    this.handleKanbanChange = this.handleKanbanChange.bind(this)
   }
 
   initState() {
@@ -69,34 +68,37 @@ export default class InnovationCICD extends React.Component {
       issueTypeSelected: [],
       codeRepoSelected: [],
       kanbanOnly: false,
-    };
+    }
   }
 
   async requiredEventTypesMissing() {
     // check if JIRAEvent and BitbucketEvent exist in the selected account
     let eventTypesList = await NerdGraphQuery.query({
-      query: gqlQuery.runQuery(this.state.accountId, 'show eventTypes'),
-    });
+      query: gqlQuery.runQuery(
+        this.state.accountId,
+        'show eventTypes since 10 weeks ago'
+      ),
+    })
     eventTypesList =
       (
         ((((eventTypesList || {}).data || {}).actor || {}).account || {})
           .nrql || {}
-      ).results || {};
+      ).results || {}
     if (
       eventTypesList.length &&
       eventTypesList.find(e => e.eventType === 'JIRAEvent') &&
       eventTypesList.find(e => e.eventType === 'BitbucketEvent')
     ) {
-      this.setState({ eventMissing: false });
-      return false;
+      this.setState({ eventMissing: false })
+      return false
     } else {
-      this.setState({ eventMissing: true });
-      return true;
+      this.setState({ eventMissing: true })
+      return true
     }
   }
 
   setApplication(inProjectName, inSprintName) {
-    this.setState({ projectName: inProjectName, sprintName: inSprintName });
+    this.setState({ projectName: inProjectName, sprintName: inSprintName })
   }
 
   componentDidMount() {
@@ -107,10 +109,10 @@ export default class InnovationCICD extends React.Component {
         nerdlet.ACCOUNT_PICKER_VALUE.CROSS_ACCOUNT,
         ...nerdlet.ACCOUNT_PICKER_DEFAULT_VALUES,
       ],
-    });
+    })
     this.setState({ accountId: this.context.accountId }, async () => {
-      await this.fetchNewRelicData();
-    });
+      await this.fetchNewRelicData()
+    })
   }
 
   componentDidUpdate() {
@@ -121,107 +123,109 @@ export default class InnovationCICD extends React.Component {
           ...this.initState(),
         },
         async () => {
-          await this.fetchNewRelicData();
+          await this.fetchNewRelicData()
         }
-      );
+      )
     }
   }
 
-  async fetchNewRelicData() {
-    let eventTypeMissingFlag = await this.requiredEventTypesMissing();
-    if (!eventTypeMissingFlag) {
-      const projectNRQL = `SELECT uniqueCount(sprintName) FROM JIRAEvent where projectName NOT LIKE '%Kanban%' FACET projectName SINCE 10 weeks ago limit MAX`;
-      const kanbanNRQL = `SELECT uniqueCount(sprintName) FROM JIRAEvent since 3 months ago limit MAX where projectName LIKE '%Kanban%' facet projectName`;
+  getProjectList = async query => {
+    const projectList = []
 
-      var projectList = [];
+    const { data } = await NrqlQuery.query({
+      accountId: this.state.accountId,
+      query,
+    })
+
+    data.forEach(result => {
+      const name = result.metadata.name
+      if (name !== 'Daylight Saving Time') {
+        projectList.push({
+          name,
+          sprints: result.data[0].sprintName,
+        })
+      }
+    })
+
+    return projectList
+  }
+
+  async fetchNewRelicData() {
+    let eventTypeMissingFlag = await this.requiredEventTypesMissing()
+    if (!eventTypeMissingFlag) {
+      const projectNRQL = `SELECT uniqueCount(sprintName) FROM JIRAEvent where projectName NOT LIKE '%Kanban%' FACET projectName SINCE 10 weeks ago limit MAX`
+      const kanbanNRQL = `SELECT uniqueCount(sprintName) FROM JIRAEvent since 3 months ago limit MAX where projectName LIKE '%Kanban%' facet projectName`
+
+      let projectList = []
 
       if (this.state.kanbanOnly) {
-        let results = await NrqlQuery.query({
-          accountId: this.state.accountId,
-          query: kanbanNRQL,
-        });
-        results.data.chart.forEach(function(result) {
-          projectList.push({
-            name: result.metadata.name,
-            sprints: result.data[0].sprintName,
-          });
-        });
+        projectList = await this.getProjectList(kanbanNRQL)
       } else {
-        let results = await NrqlQuery.query({
-          accountId: this.state.accountId,
-          query: projectNRQL,
-        });
-        results.data.chart.forEach(function(result) {
-          projectList.push({
-            name: result.metadata.name,
-            sprints: result.data[0].sprintName,
-          });
-        });
+        projectList = await this.getProjectList(projectNRQL)
       }
 
-      projectList.sort(function(a, b) {
-        var nameA = a.name.toLowerCase(),
-          nameB = b.name.toLowerCase();
+      projectList.sort((a, b) => {
+        const nameA = a.name.toLowerCase()
+        const nameB = b.name.toLowerCase()
         if (nameA < nameB)
           //sort string ascending
-          return -1;
-        if (nameA > nameB) return 1;
-        return 0; //default return value (no sorting)
-      });
+          return -1
+        if (nameA > nameB) return 1
+        return 0 //default return value (no sorting)
+      })
 
-      this.setState({ projectList, loading: false });
+      this.setState({ projectList, loading: false })
     }
   }
 
   async fetchIssueData() {
-    const { projectName, issueTypeSelected } = this.state;
-    var sprintList = [];
+    const { projectName, issueTypeSelected } = this.state
+    var sprintList = []
 
     let sprintListResults = await NerdGraphQuery.query({
       query: gqlQuery.runQuery(
         this.state.accountId,
         gqlQuery.sprintList(projectName)
       ),
-    });
+    })
     sprintListResults =
       (
         ((((sprintListResults || {}).data || {}).actor || {}).account || {})
           .nrql || {}
-      ).results || {};
+      ).results || {}
 
     sprintListResults.forEach(function(result) {
-      sprintList.push({ name: result.sprintName });
-    });
+      sprintList.push({ name: result.sprintName })
+    })
 
-    var reA = /[^a-zA-Z]/g;
-    var reN = /[^0-9]/g;
+    var reA = /[^a-zA-Z]/g
+    var reN = /[^0-9]/g
 
     function sortAlphaNum(a, b) {
-      var aA = a['name'].replace(reA, '');
-      var bA = b['name'].replace(reA, '');
+      var aA = a['name'].replace(reA, '')
+      var bA = b['name'].replace(reA, '')
       if (aA === bA) {
-        var aN = parseInt(a['name'].replace(reN, ''), 10);
-        var bN = parseInt(b['name'].replace(reN, ''), 10);
-        return aN === bN ? 0 : aN > bN ? 1 : -1;
+        var aN = parseInt(a['name'].replace(reN, ''), 10)
+        var bN = parseInt(b['name'].replace(reN, ''), 10)
+        return aN === bN ? 0 : aN > bN ? 1 : -1
       } else {
-        return aA > bA ? 1 : -1;
+        return aA > bA ? 1 : -1
       }
     }
-    sprintList.sort(sortAlphaNum);
-    sprintList = sprintList.reverse();
+    sprintList.sort(sortAlphaNum)
+    sprintList = sprintList.reverse()
 
-    this.setState({ sprintList });
+    this.setState({ sprintList })
   }
 
-  handleProjectChange(ent, value) {
-    const { projectName, sprintName } = this.state;
-    this.setState({ loading: true, sprintName: null, projectName: value });
+  handleProjectChange(value) {
+    this.setState({ loading: true, sprintName: null, projectName: value })
 
     if (value == 'All') {
       this.setState({
         projectName: null,
         loading: false,
-      });
+      })
     } else {
       this.setState(
         {
@@ -229,24 +233,22 @@ export default class InnovationCICD extends React.Component {
           sprintName: 'All',
         },
         async () => {
-          await this.fetchMappingStorage(value);
-          await this.fetchIssueData();
+          await this.fetchMappingStorage(value)
+          await this.fetchIssueData()
           this.setState({
             loading: false,
-          });
+          })
         }
-      );
+      )
     }
   }
 
   async fetchMappingStorage(value) {
-    let mappingStorage = (
-      await UserStorageQuery.query({
-        accountId: this.state.accountId,
-        collection: 'InnovationCICD',
-        documentId: value,
-      })
-    ).data;
+    const mappingStorage = await readUserCollection(
+      NerdstoreDefaults.CONFIG_COLLECTION_NAME,
+      value
+    )
+
     this.setState({
       codeRepoSelected:
         mappingStorage && mappingStorage.codeRepoSelected
@@ -256,58 +258,63 @@ export default class InnovationCICD extends React.Component {
         mappingStorage && mappingStorage.issueTypeSelected
           ? mappingStorage.issueTypeSelected
           : ['Story', 'Task'],
-    });
+      workloadSelected:
+        mappingStorage && mappingStorage.guid ? mappingStorage.guid : null,
+    })
   }
 
-  handleSettingsChange(issueTypeSelected, codeRepoSelected) {
+  handleSettingsChange(issueTypeSelected, codeRepoSelected, workloadSelected) {
     this.setState(
       {
         loading: true,
         sprintName: 'All',
       },
       () => {
-        issueTypeSelected = issueTypeSelected ? issueTypeSelected : [];
-        codeRepoSelected = codeRepoSelected ? codeRepoSelected : [];
+        issueTypeSelected = issueTypeSelected ? issueTypeSelected : []
+        codeRepoSelected = codeRepoSelected ? codeRepoSelected : []
+        workloadSelected = workloadSelected && workloadSelected
         this.setState(
           {
             issueTypeSelected,
             codeRepoSelected,
+            workloadSelected,
           },
           async () => {
-            await this.fetchIssueData();
+            await this.fetchIssueData()
             this.setState({
               loading: false,
-            });
+            })
           }
-        );
+        )
       }
-    );
+    )
   }
 
   handleKanbanChange(event) {
     this.setState(
       {
         kanbanOnly: !this.state.kanbanOnly,
+        loading: true,
         sprintName: null,
         projectName: null,
       },
       async () => {
-        await this.fetchNewRelicData();
+        await this.fetchNewRelicData()
       }
-    );
+    )
   }
 
-  async handleSprintChange(ent, value) {
-    const { projectName, sprintName } = this.state;
+  handleSprintChange(value) {
+    const { projectName, sprintName } = this.state
 
     if (value == 'All') {
       this.setState({
         sprintName: 'All',
-      });
+      })
     } else {
       this.setState({
         sprintName: value,
-      });
+      })
     }
   }
 
@@ -322,136 +329,108 @@ export default class InnovationCICD extends React.Component {
       loading,
       issueTypeSelected,
       codeRepoSelected,
+      workloadSelected,
       kanbanOnly,
-    } = this.state;
+    } = this.state
 
     return (
       <>
-        <Stack
-          className="toolbar-container"
-          fullWidth
-          gapType={Stack.GAP_TYPE.NONE}
-          horizontalType={Stack.HORIZONTAL_TYPE.FILL}
-          verticalType={Stack.VERTICAL_TYPE.CENTER}
-        >
-          <StackItem className="toolbar-section1">
-            <Stack
-              gapType={Stack.GAP_TYPE.LARGE}
-              fullWidth
-              verticalType={Stack.VERTICAL_TYPE.FILL}
-            >
-              <StackItem className="toolbar-item has-separator">
-                {projectList ? (
-                  <ProjectDropdown
-                    projectName={projectName}
-                    loadedProjects={projectList}
-                    loading={loading}
-                    filterChange={(ent, value) =>
-                      this.handleProjectChange(ent, value)
-                    }
-                  />
-                ) : (
-                  <Spinner />
-                )}
-              </StackItem>
-              <StackItem className="toolbar-item has-separator">
-                <Radio
-                  label="Kanban Only (BETA)"
-                  checked={kanbanOnly}
-                  onClick={this.handleKanbanChange}
+        <Layout style={{ marginBottom: '0px' }}>
+          <LayoutItem className="toolbar__container">
+            <div className="toolbar__section">
+              <div className="toolbar__item">
+                <SegmentedControl
+                  className="toolbar__item-element"
+                  onChange={this.handleKanbanChange}
+                  value={kanbanOnly ? 'kanban' : 'scrum'}
+                >
+                  <SegmentedControlItem value="scrum" label="Scrum" />
+                  <SegmentedControlItem value="kanban" label="Kanban" />
+                </SegmentedControl>
+              </div>
+              {projectList ? (
+                <ProjectDropdown
+                  selectedProject={projectName}
+                  loadedProjects={projectList}
+                  filterChange={this.handleProjectChange}
                 />
-              </StackItem>
-            </Stack>
-          </StackItem>
-          <Stack
-            fullWidth
-            fullHeight
-            verticalType={Stack.VERTICAL_TYPE.CENTER}
-            horizontalType={Stack.HORIZONTAL_TYPE.CENTER}
-          >
-            <StackItem className="toolbar-section2">
-              <Stack
-                fullWidth
-                fullHeight
-                verticalType={Stack.VERTICAL_TYPE.CENTER}
-                horizontalType={Stack.HORIZONTAL_TYPE.CENTER}
-              >
-                <StackItem className="toolbar-item has-separator">
-                  <HeadingText type={HeadingText.TYPE.HEADING_3}>
-                    {projectName ? projectName : 'All'}
-                  </HeadingText>
-                </StackItem>
-                <StackItem>
-                  <HeadingText type={HeadingText.TYPE.HEADING_3}>
-                    {sprintName ? sprintName : ''}
-                  </HeadingText>
-                </StackItem>
-              </Stack>
-            </StackItem>
-          </Stack>
-          {projectName && projectName != 'All' ? (
-            <Mapping
-              projectName={projectName}
-              loading={loading}
-              accountId={accountId}
-              settingsChange={(issueTypeSelected, codeRepoSelected) =>
-                this.handleSettingsChange(issueTypeSelected, codeRepoSelected)
-              }
-            />
-          ) : (
-            <></>
-          )}
-        </Stack>
-
-        <Grid
-          className="primary-grid"
-          spacingType={[Grid.SPACING_TYPE.NONE, Grid.SPACING_TYPE.NONE]}
-        >
-          <GridItem className="sidebar-container" columnSpan={2}>
-            {projectName && sprintName && !kanbanOnly ? (
-              <SprintList
+              ) : (
+                <Spinner />
+              )}
+            </div>
+            <div className="toolbar__section">
+              <div className="toolbar__item">
+                {projectName && projectName != 'All' && (
+                  <div className="button__row">
+                    <Button
+                      onClick={() =>
+                        navigation.openStackedEntity(workloadSelected)
+                      }
+                      type={Button.TYPE.PRIMARY}
+                      iconType={
+                        Button.ICON_TYPE
+                          .HARDWARE_AND_SOFTWARE__SOFTWARE__LIVE_VIEW
+                      }
+                      disabled={!workloadSelected}
+                    >
+                      View Workload
+                    </Button>
+                    <ConfigModal
+                      projectName={projectName}
+                      loading={loading}
+                      accountId={accountId}
+                      settingsChange={this.handleSettingsChange}
+                    />
+                  </div>
+                )}
+              </div>
+            </div>
+          </LayoutItem>
+        </Layout>
+        <Layout fullHeight>
+          {projectName && (
+            <LayoutItem
+              type={LayoutItem.TYPE.SPLIT_LEFT}
+              sizeType={LayoutItem.SIZE_TYPE.SMALL}
+              className="navigation__container"
+            >
+              <Navigation
                 projectName={projectName}
                 loadedSprints={sprintList}
                 sprintName={sprintName}
-                filterChange={(ent, value) =>
-                  this.handleSprintChange(ent, value)
-                }
+                onChange={this.handleSprintChange}
               />
-            ) : (
-              <HeadingText type={HeadingText.TYPE.HEADING_4}>
-                <center></center>
-              </HeadingText>
-            )}
-          </GridItem>
-
-          <GridItem className="primary-content-container" columnSpan={10}>
-            <main className="primary-content full-height">
-              {eventMissing ? (
+            </LayoutItem>
+          )}
+          <LayoutItem className="dashboard__container">
+            {eventMissing ? (
+              <center>
+                <h3>
+                  No "JIRAEvent" and/or "BitbucketEvent" Types found in this
+                  account
+                </h3>
+                <h4>Please select an account with required event types</h4>
+              </center>
+            ) : loading ? (
+              <>
+                <Spinner /> <br />
                 <center>
-                  <h3>No "JIRAEvent" and/or "BitbucketEvent" Types found in this account</h3>
-                  <h4>Please select an account with required event types</h4>
+                  {'Loading Types: ' + issueTypeSelected.toString()}
                 </center>
-              ) : loading ? (
-                <>
-                  <Spinner /> <br />
-                  <center>
-                    {'Loading Types: ' + issueTypeSelected.toString()}
-                  </center>
-                </>
-              ) : (
-                <Dashboard
-                  projectName={projectName}
-                  accountId={accountId}
-                  kanbanOnly={kanbanOnly}
-                  sprintName={sprintName}
-                  issueTypeSelected={issueTypeSelected}
-                  codeRepoSelected={codeRepoSelected}
-                />
-              )}
-            </main>
-          </GridItem>
-        </Grid>
+              </>
+            ) : (
+              <Dashboard
+                projectName={projectName}
+                accountId={accountId}
+                kanbanOnly={kanbanOnly}
+                sprintName={sprintName}
+                issueTypeSelected={issueTypeSelected}
+                codeRepoSelected={codeRepoSelected}
+              />
+            )}
+          </LayoutItem>
+        </Layout>
       </>
-    );
+    )
   }
 }
